Move bitlist and bitvector into own files
